requirements.txt
streamlit>=1.33.0
openai>=1.35.0
pypdf>=4.2.0
python-docx>=1.1.0
tenacity>=8.3.0
cohere>=5.5.8 